Split-out driver config.
